// web/src/api/generated/model/card.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Campus Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface Card {
  name: string;
  student_number: string;
  barcode: string;
  programme_name: string;
  class_group: string | null;
  valid_until: string;
  has_photo: boolean;
}